Enables mixed precision training with AMP
Implements Automatic Mixed Precision (AMP) training to improve performance and reduce memory usage.

- Adds a `use_amp` flag to the configuration file to enable/disable AMP.
- Introduces `GradScaler` for gradient scaling during AMP training.
- Uses `autocast` context manager to perform operations in mixed precision.
- Modifies forward passes to utilize `autocast` when `use_amp` is enabled.
- Casts K_s to float before inverse operation
- Compiles the depth and pose networks for faster execution.
- Adds `torch.cuda.empty_cache()` and `gc.collect()` to reduce memory fragmentation and prevent OOM errors.

+++ vo/monodepth_learner_torch.py
@@ -179,7 +179,7 @@ class MonodepthLearner:
                 target = tgt_image
                 K_s = intrinsic
             
-            inv_K_s = torch.inverse(K_s)
+            inv_K_s = torch.inverse(K_s.float()).type(K_s.dtype)
             
             # Compute reprojection for each source
             reprojection_losses = []
@@ -332,7 +332,8 @@ class MonodepthLearner:
             T = T_4x4[:, :3, :]
             
             # Warp source to target
-            inv_K_s = torch.inverse(K_s)
+            inv_K_s = torch.inverse(K_s.float()).type(K_s.dtype)
+            
             cam_points = self.backproject_depth[scale](depth, inv_K_s)
             pix_coords = self.project_3d[scale](cam_points, K_s, T)
